+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.39](https://github.com/arez/arez/tree/v0.39) (2018-01-09)
 [Full Changelog](https://github.com/arez/arez/compare/v0.38...v0.39)