+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+#### Changed
+* 💥 **\[processor\]** Stop the annotation processor generating the `[Name]BaseRepositoryExtension`
+  interface when generating repositories as it offered limited benefit relative to the complexity
+  it introduced.
+
 ### [v0.39](https://github.com/arez/arez/tree/v0.39) (2018-01-09)
 [Full Changelog](https://github.com/arez/arez/compare/v0.38...v0.39)